Moved EntityIndex and PrimaryIndex to separate files
