+++ Entitas/Entitas/Index/IEntityIndex.cs
@@ -1,0 +1,7 @@
+﻿namespace Entitas {
+
+    public interface IEntityIndex {
+        void Activate();
+        void Deactivate();
+    }
+}
